+++ src/config/__tests__/app.test.ts
@@ -24,7 +24,9 @@ describe('getServerConfig', () => {
   describe('index url', () => {
     it('should return default URLs when no environment variables are set', () => {
       const config = getAppConfig();
-      expect(config.AGENTS_INDEX_URL).toBe('https://registry.npmmirror.com/@lobehub/agents-index/v1/files/public');
+      expect(config.AGENTS_INDEX_URL).toBe(
+        'https://registry.npmmirror.com/@lobehub/agents-index/v1/files/public',
+      );
       expect(config.PLUGINS_INDEX_URL).toBe('https://chat-plugins.lobehub.com');
     });
 
@@ -41,7 +43,9 @@ describe('getServerConfig', () => {
       process.env.PLUGINS_INDEX_URL = '';
 
       const config = getAppConfig();
-      expect(config.AGENTS_INDEX_URL).toBe('https://registry.npmmirror.com/@lobehub/agents-index/v1/files/public');
+      expect(config.AGENTS_INDEX_URL).toBe(
+        'https://registry.npmmirror.com/@lobehub/agents-index/v1/files/public',
+      );
       expect(config.PLUGINS_INDEX_URL).toBe('https://chat-plugins.lobehub.com');
     });
   });