+++ src/services/textToImage.ts
@@ -9,7 +9,10 @@ interface FetchOptions {
 }
 
 class ImageGenerationService {
-  async generateImage(params: Omit<OpenAIImagePayload, 'model' | 'n'>, options?: FetchOptions) {
+  generateImage = async (
+    params: Omit<OpenAIImagePayload, 'model' | 'n'>,
+    options?: FetchOptions,
+  ) => {
     const payload: OpenAIImagePayload = { ...params, model: 'dall-e-3', n: 1 };
 
     const provider = ModelProvider.OpenAI;
@@ -32,7 +35,7 @@ class ImageGenerationService {
     const urls = await res.json();
 
     return urls[0] as string;
-  }
+  };
 }
 
 export const imageGenerationService = new ImageGenerationService();